Duplicate Button, Typing Descriptions
Pressing Duplicate Graph does not send the user to a separate screen; the interaction generates a pop up that asks the user to confirm whether they would like to duplicate the current graph or not. Provides a confirmation once duplicated. Currently calls on a function in aysncStorage called dupeGraph() that does not actually duplicate the graph (work in progress).

Added KeyboardAvoidingView to the list of imports from React Native

While adding descriptions to a data point, the screen will automatically follow what the user is currently typing, allowing the user to see what they most recently typed to avoid the keyboard blocking the text.